% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow.R
\docType{data}
\name{COM44}
\alias{COM44}
\title{COM44}
\description{
SpatialPolygonsDataFrame of Loire Atlantique (communal level) - same variables than \link{MRE44}
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3}
}

