% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{commuters_datasets}
\alias{commuters_datasets}
\title{Commuters datasets}
\description{
Data on commuters between Urban Areas of the French Grand Est region in 2011.
Fields: \cr
\itemize{
\item{i: Code of the urban area of residence}
\item{namei: Name of the urban area of residence}
\item{wi: Total number of active occupied persons in the urban area of residence}
\item{j: Code of the urban area of work}
\item{namej: Name of the urban area of work}
\item{wj: Total number of active occupied persons in the urban area of work}
\item{fij: Number of commuters between i and j}
}

Geopackage of the Grand Est region in France and its urban areas (2010 delineation).
}
\examples{
nav <- read.csv(system.file("csv/nav.csv", package = "flows"))
library(sf)
UA <- st_read(system.file("gpkg/GE.gpkg", package = "flows"), layer = "urban_area")
GE <- st_read(system.file("gpkg/GE.gpkg", package = "flows"), layer = "region")
}
\references{
Commuters dataset: \url{https://www.insee.fr/fr/statistiques/2022113} \cr
Spatial dataset: \url{https://geoservices.ign.fr/geofla}
}
