% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{compmat}
\alias{compmat}
\title{Comparison of Two Matrices}
\usage{
compmat(mat1, mat2, digits = 0)
}
\arguments{
\item{mat1}{A square matrix of flows.}

\item{mat2}{A square matrix of flows.}

\item{digits}{An integer indicating the number of decimal places to be used
when printing the data.frame in the console (see \link{round}).}
}
\value{
A data.frame that provides statistics on differences
between mat1 and mat2: absdiff are the
absolute diffences and reldiff are the relative differences (in percent).
}
\description{
Compares two matrices of same dimension, with same column and
row names.
}
\examples{
data(nav)
mat <- prepflows(mat = nav, i = "i", j = "j", fij = "fij")
diag(mat) <- 0
x <- domflows(mat = mat, wi = colSums(mat), wj = colSums(mat), k = 1)
firstx <- firstflows(mat = mat, method = "nfirst", ties.method = "first", k = 1)
xnb <- firstflows(mat = mat, method = "xfirst", ties.method = "first", k = 20)
mat2 <- mat * firstx * x * xnb
compmat(mat1 = mat, mat2 = mat2, digits = 1)
}
\seealso{
\link{statmat}
}

