% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{domflows}
\alias{domflows}
\title{Dominant Flows Selection}
\usage{
domflows(mat, wi, wj, k)
}
\arguments{
\item{mat}{A square matrix of flows}

\item{wi}{A vector of weight for i}

\item{wj}{A vector of weight for j}

\item{k}{Threshold. wj/wi> k}
}
\description{
Compute the dominant flow analysis
}

