% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{firstflows}
\alias{firstflows}
\title{Flow Selection from Origins}
\usage{
firstflows(mat, method = "nfirst", ties.method = "first", k)
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{method}{One of "nfirst", "xfirst" or "xsumfirst":
\cr nfirst selects k first flows from origins,
\cr xfirst selects flows from origins greater than k,
\cr xsumfirst selects flows from origins while sum of flows is smaller than k.}

\item{ties.method}{In case of equality with "nfirst" method (use "random" or "first", see \link{rank}).}

\item{k}{Selection threshold.}
}
\value{
A boolean matrix of selected flows.
}
\description{
Flow selection from origins.
}
\details{
As the output is a boolean matrix, use element-wise multiplication to get flows intensity.
}
\examples{
data(nav)
myflows <- prepflows(mat = nav, i = "i", j = "j", fij = "fij")

#remove diagonal
diag(myflows) <- 0
statmat(myflows)

#select 2 flows per spatial unit
fflows1 <- firstflows(myflows, method = "nfirst", ties.method = "first", k = 2)
# Use element-wise multiplication to get flows intensity.
fflow1 <- fflows1 * myflows
statmat(fflow1)

#select flows > 20
fflows2 <- firstflows(myflows, method = "xfirst", k = 20)
fflow2 <- fflows2 * myflows
statmat(fflow2)

# Select each flows that represent at least 10\% of the outputs
myflowspct <- myflows / rowSums(myflows) * 100
fflows2 <- firstflows(mat = myflowspct, method = "xfirst", k = 10)
fflow2 <- fflows2 * myflows
statmat(fflow2)
}
\seealso{
\link{firstflowsg}, \link{domflows}
}

