% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{firstflowsg}
\alias{firstflowsg}
\title{Flow Selection from Full Matrix}
\usage{
firstflowsg(mat, method = "nfirst", k, ties.method = "first")
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{method}{A method of flow selection, one of "nfirst", "xfirst" or "xsumfirst":
\itemize{
\item{nfirst selects k first flows,}
\item{xfirst selects flows greater than k,}
\item{xsumfirst selects as many flows necessary so that their sum is at least equal to k.}
}}

\item{k}{Selection threshold.}

\item{ties.method}{In case of equality with "nfirst" method (use "random" or "first", see \link{rank}).}
}
\value{
A boolean matrix of selected flows.
}
\description{
Flow selection based on global criteria.
}
\details{
As the output is a boolean matrix, use element-wise multiplication to get flows intensity.
}
\examples{
data(nav)
myflows <- prepflows(mat = nav, i = "i", j = "j", fij = "fij")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#select 50 first flow on the whole matric
 xx <- firstflowsg(myflows, method = "nfirst", ties.method = "first", 50)
fflow1 <- fflows1 * myflows
statmat(fflow1)
#select flows > 50
fflows2 <- firstflowsg(myflows, method = "xfirst", ties.method = "first", 50)
fflow2 <- fflows2 * myflows
statmat(fflow2)

# Select flows that represent at least 50\% of the matrix flows
k50 <- sum(myflows)/2
fflows2 <- firstflowsg(myflows, method = "xsumfirst", k = 150000)
}
\seealso{
\link{firstflows}, \link{domflows}
}

