% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{flowDom}
\alias{flowDom}
\title{Compute dominant flows}
\usage{
flowDom(mat, i, j, fij, k = NULL)
}
\arguments{
\item{mat}{A data.frame of flows between \code{i} and \code{j} (long format).}

\item{i}{A character giving the name of emiting units field in \code{mat}.}

\item{j}{A character giving the name of receiving units field in \code{mat}.}

\item{fij}{A character giving the name of the flow field in \code{mat}.}

\item{k}{A numeric giving the minimum share of flow that must be emitted from
\code{i} to \code{j}. (optional)}
}
\description{
Compute dominant flows on a long matrix (i, j, fij). Two
methods are allowed: i sends its major flow to j or i sends a specific share
to j.
}
\details{
The output of the function is a data.frame giving couples where
\code{i} is dominated by \code{j} and the share of outputs from i to j in the
total outputs from i.
}
\examples{
data(lot)
dom1<- flowDom(mat = MRE46,
              i = "DCRAN",
              j = "CODGEO",
              fij = "NBFLUX_C08_POP05P")
dom2 <- flowDom(mat = MRE46,
               i = "DCRAN",
               j = "CODGEO",
               fij = "NBFLUX_C08_POP05P",
               k = 0.5)
}
\references{
Nystuen, J. D. and Dacey, M. F. (1961), A GRAPH THEORY
INTERPRETATION OF NODAL REGIONS. Papers in Regional Science, 7: 29-42.
}

