% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{flowPrep}
\alias{flowPrep}
\title{Flows Preparation}
\usage{
flowPrep(mat, i, j, fij)
}
\arguments{
\item{mat}{A data.frame of flows between \code{i} and \code{j} (long format
matrix: \code{(i, j, fij)}.}

\item{i}{A character giving the origin field in \code{mat}.}

\item{j}{A character giving the destination field in \code{mat}.}

\item{fij}{A character giving the flow field in \code{mat}.}
}
\description{
From a long format matrix to a weight data.frame and a wide
squared matrix
}
\examples{
data(LoireAtlantique)
}

