% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{flows-package}
\alias{flows-package}
\title{Package description}
\description{
Selections on flow matrices, statistics on selected flows, map
and graph visualisations.

An introduction to the package conceptual background and usage is proposed in
a vignette (see \code{vignette(topic = "flows")}) and a paper (Beauguitte, Giraud & Guérois 2015).
}
\references{
L. Beauguitte, T. Giraud & M. Guérois, 2015. "Un outil pour la sélection et la visualisation de flux : le package flows",
 \emph{Netcom}, 29-3/4:399-408. \url{https://journals.openedition.org/netcom/2134}.
}
