% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{GE}
\alias{GE}
\title{Grand Est Region}
\description{
SpatialPolygonsDataFrame of the Grand Est region in France.
}
\examples{
###GE
data(nav)
sp::plot(GE, col = "#cceae7", border = "grey50")
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3}
}

