% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow.R
\docType{data}
\name{MRE44}
\alias{MRE44}
\title{MRE44}
\description{
Migration flows at the communal level for departement of Loire Atlantique (change of residence between 2008 and 2012)
}
\details{
Variables: \cr
\itemize{
\item{CODGEO: Departement and commune of residence}
\item{LIBGEO: Name of the commune of residence}
\item{DCRAN: INSEE code of former departement and commune of residence (municipal arrondissement for Paris, Lyon, Marseille)}
\item{L_DCRAN: Name of former commune (municipal arrondissement for Paris, Lyon, Marseille)}
\item{NBFLUX_C08_POP05P: Number of migrants (5 years or more) from commune or arrondissement DCRAN to commune CODGEO}}
}
\references{
\url{http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=migration-residentielle-08}
}

