% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{nav}
\alias{nav}
\title{Commuters}
\description{
Data on commuters between Urban Areas of the French Grand Est region in 2011.
Fields: \cr
\itemize{
\item{i: Code of the urban area of residence}
\item{namei: Name of the urban area of residence}
\item{wi: Total number of active occupied persons in the urban area of residence}
\item{j: Code of the urban area of work}
\item{namej: Name of the urban area of work}
\item{wj: Total number of active occupied persons in the urban area of work}
\item{fij: Number of commuters between i and j}
}
}
\examples{
## nav
data(nav)
str(nav)
}
\references{
\url{http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=mobilite-professionnelle-11}
}
