% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_flows.R
\name{nodal_flows}
\alias{nodal_flows}
\title{Nodal flows selection}
\usage{
nodal_flows(mat)
}
\arguments{
\item{mat}{A square matrix of flows.}
}
\value{
The matrix of the selected flows is returned.
}
\description{
Perform a Nystuen & Dacey's dominants flows analysis.
}
\examples{
nav <- read.csv(system.file("csv/nav.csv", package = "flows"))
mat <- prepare_mat(x = nav, i = "i", j = "j", fij = "fij")
res <- nodal_flows(mat)
res[1:5,1:5]
}
