% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{nodalflows}
\alias{nodalflows}
\title{Nodal Flows}
\usage{
nodalflows(mat)
}
\arguments{
\item{mat}{A square matrix of flows.}
}
\description{
This function computes nodal flows following the Nystuen & Dacey method.
}
\examples{
# Import data
}
\references{
J. Nystuen & M. Dacey, 1961, "A Graph Theory Interpretation of Nodal Regions",
\emph{Papers and Proceedings of the Regional Science Association}, 7:29-42.
}
