% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodal_flows.R
\name{plot_nodal_flows}
\alias{plot_nodal_flows}
\title{Nodal flows graph}
\usage{
plot_nodal_flows(
  mat,
  leg_pos_flows = "topright",
  leg_flow = "Flows Intensity",
  leg_pos_node = "bottomright",
  leg_node = c("Dominant", "Intermediary", "Dominated",
    "Size proportional\\nto sum of inflows"),
  labels = FALSE
)
}
\arguments{
\item{mat}{A square matrix of dominant flows (see \link{nodal_flows}).}

\item{leg_pos_flows}{Position of the flows legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{leg_flow}{Title of the flows legend.}

\item{leg_pos_node}{Position of the nodes legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{leg_node}{Text of the nodes legend.}

\item{labels}{A boolean, if TRUE, labels of dominant and intermediary nodes are plotted.}
}
\description{
This function plots a dominant flows graph.
}
\details{
This function uses the Davidson Harel algorithm from igraph.
}
\note{
As square matrices can easily be plotted with \link[igraph]{plot.igraph} or
\link[sna]{gplot} functions from igraph and sna packages, we do not propose
visualisation for other outputs.
}
\examples{
nav <- read.csv(system.file("csv/nav.csv", package = "flows"))
mat <- prepare_mat(x = nav, i = "i", j = "j", fij = "fij")
res <- nodal_flows(mat)

# Plot dominant flows graph
plot_nodal_flows(mat = res)
}
