% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{plotDomFlows}
\alias{plotDomFlows}
\title{Dominant Flows Graph}
\usage{
plotDomFlows(mat)
}
\arguments{
\item{mat}{A square matrix of flows}
}
\description{
Display a dominant flows graph
}
\note{
We do not propose visualisation for other outputs as square matrices produced can easily be plot
(after transformation) with \link[igraph]{plot.igraph} or \link[sna]{gplot} functions.
}
\examples{
data(LoireAtlantique)
mat <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
diag(mat) <- 0
x <- domflows(mat = mat, wi = colSums(mat), wj = colSums(mat), k = 1)
firstx <- firstflows(mat = mat, method = "nfirst", ties.method = "first", k = 1)
xnb <- firstflows(mat = mat, method = "xfirst", ties.method = "first", k = 20)
mat <- mat * firstx * x * xnb
plotDomFlows(mat)
}

