% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{plotflowDom}
\alias{plotflowDom}
\title{Plot dominant flows}
\usage{
plotflowDom(fdom, spdf, id, name = NULL)
}
\arguments{
\item{fdom}{A data.frame outputed by the flowDom function.}

\item{spdf}{A SpatialPolygonsDataFrame to be linked to fdom.}

\item{id}{A character giving the name of the 'id' field in spdf.}

\item{name}{A character giving the name of a label field in spdf to be
plotted. (optional)}
}
\description{
Plot a map of the Dominant Flows. It uses, as input, the output
of the flowDom function.
}
\details{
The output of the function is a plot of a map. The map shows which
units are either "dominant", "dominated" or both. The darker a link is the
higher the share of the flow is in the total flows emitted by the unit.
}
\examples{
data(lot)
dom1<- flowDom(mat = MRE46,
              i = "DCRAN",
              j = "CODGEO",
              fij = "NBFLUX_C08_POP05P")
dom2 <- flowDom(mat = MRE46,
               i = "DCRAN",
               j = "CODGEO",
               fij = "NBFLUX_C08_POP05P",
               k = 0.5)
plotflowDom(fdom = dom1, spdf = COM46, id = "INSEE_COM", name = "NOM_COM")
}

