% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{plotNodalFlows}
\alias{plotNodalFlows}
\title{Plot Nodal Flows}
\usage{
plotNodalFlows(mat, spdf, spdfid, cex = 0.04,
  legend.flows.pos = "topright", legend.flows.title = "flow intensity",
  legend.nodes.pos = "topleft", legend.node.txt = c("Dominant",
  "Intermediary", "Dominated", "Size proportional\\nto sum of inflows"),
  add = FALSE)
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{spdf}{A SpatialPolygonsDataFrame or a SpatialPointsDataFrame of units.}

\item{spdfid}{Name of the unique identifier variable in the spdf data.frame.}

\item{cex}{Share of the surface of the map occupied by circles (0.02 is 2\%).}

\item{legend.flows.pos}{Position of the flows legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{legend.flows.title}{Title of the flows legend.}

\item{legend.nodes.pos}{Position of the nodes legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{legend.node.txt}{Text of the nodes legend.}

\item{add}{A boolean, if TRUE, add the layer to an existing plot.}
}
\description{
This function plots nodal flows following the Nystuen & Dacey method.
}
\examples{
# Import data
}
\references{
J. Nystuen & M. Dacey, 1961, "A Graph Theory Interpretation of Nodal Regions",
\emph{Papers and Proceedings of the Regional Science Association}, 7:29-42.
}
