% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{prepflows}
\alias{prepflows}
\title{Flows Preparation}
\usage{
prepflows(mat, i, j, fij)
}
\arguments{
\item{mat}{A data.frame of flows between origins and destinations: long format
matrix (origins, destinations, flows).}

\item{i}{A character giving the origin field name in \code{mat}.}

\item{j}{A character giving the destination field name in \code{mat}.}

\item{fij}{A character giving the flow field name in \code{mat}.}
}
\value{
A square flows matrix.
}
\description{
From a long format matrix to a data.frame with sum of outputs and sum of inputs and a wide
matrix of flows.
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
myflows[1:5,1:5]
}

