% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_flows.R
\name{select_flows}
\alias{select_flows}
\title{Flow selection}
\usage{
select_flows(mat, method = "nfirst", ties = "first", global = FALSE, k, w)
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{method}{A method of flow selection, one of "dominant", "nfirst",
"xfirst" or "xsumfirst":
\itemize{
\item{dominant selects the dominant flows (see Details)}
\item{nfirst selects the k first flows from origins,}
\item{xfirst selects flows greater than k,}
\item{xsumfirst selects as many flows as necessary for each origin so that their sum is at least equal to k.
If k is not reached for one origin, all its flows are selected.}
}}

\item{ties}{In case of equality with "nfirst" method, use "random" or "first" (see \link{rank}).}

\item{global}{If TRUE flows selections is done at the matrix scale.}

\item{k}{Selection threshold for nfirst, xfirst and xsumfirst methods,
ratio for dominant method.}

\item{w}{A vector of units weigths (sum of incoming flows, sum of outgoing flows...).}
}
\value{
A boolean matrix of selected flows.
Use element-wise multiplication to get flows intensity.
}
\description{
Flow selection from origins.
}
\details{
If method = "dominant", select which flow (fij or fji) must be kept.
If the ratio weight of destination (wj) / weight of origin (wi) is greater
than k, then fij is selected and fji is not.
This function can perform the second criterion of the Nystuen &
Dacey's dominants flows analysis.
}
\examples{
# Import data
nav <- read.csv(system.file("csv/nav.csv", package = "flows"))
# Prepare data
mat <- prepare_mat(x = nav, i = "i", j = "j", fij = "fij")
# remove diagonal
diag(mat) <- 0

# Select the first flow from each origin
res <- select_flows(mat = mat, method = "nfirst", global = FALSE, k = 1)
rowSums(res)

# Select the 5 first flows of the matrix
res <- select_flows(mat = mat, method = "nfirst", global = TRUE, k = 5)
sum(res)

# Select the flows greater than 5000
res <- select_flows(mat = mat, method = "xfirst", k = 5000)
r <- mat * res
r[r>0]

# Select as many flows as necessary for each origin so that their sum is at least equal to 500.
res <- select_flows(mat = mat, method = "xsumfirst", global = FALSE, k = 500)
r <- mat * res
rowSums(r)

# Select as many flows in the matrix so that their sum is at least equal to 50000.
res <- select_flows(mat = mat, method = "xsumfirst", global = TRUE, k = 50000)
r <- mat * res
sum(rowSums(r))

# Select dominant flows
m <- mat[1:5,1:5]
ws <- colSums(m)
res <- select_flows(mat = m, method = "dominant", k = 1, w = ws)
# 2nd element has a lower weight than 3rd element (ratio > 1)
ws[3] / ws[2]
# The flow from 2nd element to 3rd element is kept
res[2, 3]
# The flow from 3rd element to 2nd element is removed
res[3, 2]
}
