% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{UA}
\alias{UA}
\title{Urban Areas}
\description{
SpatialPolygonsDataFrame of Urban Areas of the Grand-Est region
(2010 delineation).
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3}
}

