% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{domflows}
\alias{domflows}
\title{Dominant Flows Selection}
\usage{
domflows(mat, wi, wj, k)
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{wi}{A vector of weight for rows.}

\item{wj}{A vector of weight for columns.}

\item{k}{A threshold (see Details).}
}
\value{
A boolean matrix of selected flows.
}
\description{
Find dominant flows in a matrix.
}
\details{
This function selects which flow (fij or fji) must be kept.
If the ratio weight of destination (wj) / weight of origin (wi) is greater than k, then fij is kept.
This function can perform the second criterion of the Nystuen &
Dacey's dominants flows analysis.\cr
As the output is a boolean matrix, use element-wise multiplication to get flows intensity.
}
\examples{
data(nav)
myflows <- prepflows(mat = nav, i = "i", j = "j", fij = "fij")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#create the tree of dominant, intermediary and dominated nodes
domf <- domflows(mat = myflows, wi = colSums(myflows), wj = colSums(myflows), k = 1)
domfw <-domf * myflows
statmat(domfw)
}
\references{
J. Nystuen & M. Dacey, 1961, A graph theory interpretation of nodal flows,
\emph{Papers and Proceedings of the Regional Science Association}, vol. 7,  29-42.
}
\seealso{
\link{firstflows}, \link{firstflowsg}, \link{plotDomFlows}, \link{plotMapDomFlows}
}

