% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{domflows}
\alias{domflows}
\title{Dominant Flows Selection}
\usage{
domflows(mat, wi, wj, k)
}
\arguments{
\item{mat}{A square matrix of flows}

\item{wi}{A vector of weight for i}

\item{wj}{A vector of weight for j}

\item{k}{Threshold. wj/wi> k}
}
\value{
A boolean matrix of selected flows. To incorporate flows intensity, making the product
mat * boolean matrix is necessary.
}
\description{
Find dominant flows in a matrix.
}
\details{
This function can perform the second criterion of the Nyusten &
Dacey's dominants flows analysis.
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#create the tree of dominant, intermediary and dominated nodes
domf <- domflows(mat = myflows, wi = colSums(myflows), wj = colSums(myflows), k = 1)
domfw <-domf * myflows
statmat(domfw)
}
\references{
J. Nystuen & M. Dacey, 1961, A graph theory interpretation of nodal flows,
\emph{Papers and Proceedings of the Regional Science Association}, vol. 7,  29-42.
}

