% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow.R
\name{firstflows}
\alias{firstflows}
\title{Flow Selection From Origins}
\usage{
firstflows(mat, method = "nfirst", ties.method = "first", k)
}
\arguments{
\item{mat}{A square matrix of flows \emph{fij}}

\item{method}{One of "nfirst", "xfirst" or "xsumfirst". \cr nfirst = select \emph{k} first \emph{fij}
from \emph{i}
\cr xfirst = select x \emph{fij} from i where \emph{fij} > k  \cr xsumfirst = select x \emph{fij} from \emph{i}
 while sum(\emph{fij}) < k.}

\item{ties.method}{In case of equality with 'nfirst' method.}

\item{k}{Selection threshold (can be relative or absolute).}
}
\value{
A boolean matrix of selected flows. To incorporate flows intensity, making the product
mat * boolean matrix is necessary.
}
\description{
Flow selection from \emph{i}.
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#select 2 flows per spatial unit
fflows1 <- firstflows(myflows, method = "nfirst", ties.method = "first", 2)
fflow1 <- fflows1 * myflows
statmat(fflow1)
#select flows > 20
fflows2 <- firstflows(myflows, method = "xfirst", ties.method = "first", 20)
fflow2 <- fflows2 * myflows
statmat(fflow2)
#select sum(flows) from i > 20
fflows3 <- firstflows(myflows, method = "xsumfirst", ties.method = "first", 20)
fflow3 <- fflows3 * myflows
statmat(fflow3)
}
\references{
J. Nystuen & M. Dacey, 1961, A graph theory interpretation of nodal flows,
\emph{Papers and Proceedings of the Regional Science Association}, vol. 7,  29-42.
}

