% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{firstflows}
\alias{firstflows}
\title{Flow Selection From Origins}
\usage{
firstflows(mat, method = "nfirst", ties.method = "first", k)
}
\arguments{
\item{mat}{A square matrix of flows}

\item{method}{One of "nfirst", "xfirst" or "xsumfirst". \cr nfirst = select k first fij from i
\cr xfirst = select x fij from i where fij > k  \cr xsumfirst = select x fij from i while sum(fij) < k.}

\item{ties.method}{In case of equality with 'nfirst' method.}

\item{k}{Selection threshold.}
}
\value{
A boolean matrix of selected flows
}
\description{
Flow selection from i.
}
\references{
Nystuen & Dacey, 1961
}

