% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{firstflowsg}
\alias{firstflowsg}
\title{Flow Selection from Full Matrix}
\usage{
firstflowsg(mat, method = "nfirst", k, ties.method = "first")
}
\arguments{
\item{mat}{A square matrix of flows.}

\item{method}{One of "nfirst", "xfirst" or "xsumfirst":
\cr nfirst selects the k first flows,
\cr xfirst selects flows greater than k,
\cr xsumfirst selects flows while sum of flows is smaller than k.}

\item{k}{Selection threshold.}

\item{ties.method}{In case of equality with "nfirst" method (use "random" or "first", see \link{rank}).}
}
\value{
A boolean matrix of selected flows.
}
\description{
Flow selection based on global criteria.
}
\details{
As the output is a boolean matrix, use element-wise multiplication to get flows intensity.
}
\examples{
data(nav)
myflows <- prepflows(mat = nav, i = "i", j = "j", fij = "fij")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#select 50 first flow on the whole matric
fflows1 <- firstflowsg(myflows, method = "nfirst", ties.method = "first", 50)
fflow1 <- fflows1 * myflows
statmat(fflow1)
#select flows > 50
fflows2 <- firstflowsg(myflows, method = "xfirst", ties.method = "first", 50)
fflow2 <- fflows2 * myflows
statmat(fflow2)
}
\seealso{
\link{firstflows}, \link{domflows}
}

