% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{firstflowsg}
\alias{firstflowsg}
\title{Flow Selection From the Total Matrix}
\usage{
firstflowsg(mat, method = "nfirst", k, ties.method = "first")
}
\arguments{
\item{mat}{A matrix}

\item{method}{A method}

\item{k}{A k}

\item{ties.method}{A ties.method}
}
\description{
Various flow selection on global kriterions
}

