% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flow.R
\name{firstflowsg}
\alias{firstflowsg}
\title{Flow Selection From the Total Matrix}
\usage{
firstflowsg(mat, method = "nfirst", k, ties.method = "first")
}
\arguments{
\item{mat}{A square matrix of flows \emph{fij}}

\item{method}{One of "nfirst", "xfirst" or "xsumfirst". \cr nfirst = select \emph{k} first \emph{fij}
from the input matrix
\cr xfirst = select x \emph{fij} from the input matrix where \emph{fij} > k  \cr xsumfirst = select x \emph{fij} from the input matrix
 while sum(\emph{fij}) < k.}

\item{k}{Selection threshold (can be relative or absolute).}

\item{ties.method}{In case of equality with 'nfirst' method (see \link{rank},
"random" and "first" select only one flow, other methods may select various).}
}
\value{
A boolean matrix of selected flows. To incorporate flows intensity, making the product
mat * boolean matrix is necessary.
}
\description{
Various flow selection on global criterions
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
#remove diagonal
diag(myflows) <- 0
statmat(myflows)
#select 50 first flow on the whole matric
fflows1 <- firstflowsg(myflows, method = "nfirst", ties.method = "first", 50)
fflow1 <- fflows1 * myflows
statmat(fflow1)
#select flows > 50
fflows2 <- firstflowsg(myflows, method = "xfirst", ties.method = "first", 50)
fflow2 <- fflows2 * myflows
statmat(fflow2)
#select sum(flows) > 50\% of the whole matrix
#I don't know how to make a relative selection...
}

