% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{flowDom}
\alias{flowDom}
\title{Dominant Flows Analysis}
\usage{
flowDom(mat, i, j, fij, k = NULL)
}
\arguments{
\item{mat}{A data.frame of flows between \code{i} and \code{j} (long format
matrix: \code{(i, j, fij)}.}

\item{i}{A character giving the origin field in \code{mat}.}

\item{j}{A character giving the destination field in \code{mat}.}

\item{fij}{A character giving the flow field in \code{mat}.}

\item{k}{A numeric giving the minimum share of \code{i} outbound flows that
must be sent to \code{j}. (variant - optional)}
}
\description{
Computes dominant flows on a matrix :

\code{i} is dominated by \code{j} when \code{i} sends its biggest flow to
\code{j} and \code{j} receives more total inbound flows than \code{i}. A
variant is that \code{i} is dominated by \code{j} when \code{i} sends at
least a specific share (\code{k}) of its outbound flows to \code{j} and
\code{j} receives more total inbound flows than \code{i}.
}
\details{
The output of the function is a data.frame giving couples where
\code{i} is dominated by \code{j}. Other variables are :\itemize{
 \item{dom: }{share of \code{i} outbound flows sent to \code{j}.}
 \item{fij: }{flow from \code{i} to \code{j}}
 \item{sumInI: }{\code{i} total inbound flows}
 \item{sumOutI: }{\code{i} total outbound flows}
 \item{sumInJ: }{\code{j} total inbound flows}
 \item{sumOutJ: }{\code{j} total outbound flows}
}
}
\examples{
data(LoireAtlantique)
dom1<- flowDom(mat = MRE44,
              i = "DCRAN",
              j = "CODGEO",
              fij = "NBFLUX_C08_POP05P")
## or
dom2 <- flowDom(mat = MRE44,
               i = "DCRAN",
               j = "CODGEO",
               fij = "NBFLUX_C08_POP05P",
               k = 0.5)
}
\references{
Nystuen, J. D. and Dacey, M. F. (1961), A GRAPH THEORY
INTERPRETATION OF NODAL REGIONS. Papers in Regional Science, 7: 29-42.
}

