% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\docType{package}
\name{flows}
\alias{flows}
\alias{flows-package}
\title{Flows Selection and Analysis}
\description{
This package contains various functions to filter flows matrices
(dominant and major flows), provides statistics on selection made
 and proposes map and graph visualisation.
}

