% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\docType{package}
\name{flows}
\alias{flows}
\alias{flows-package}
\title{Flows Selection and Analysis}
\description{
Selections on flow matrices, statistics on selected flows, map
and graph visualisations.

An introduction to the package conceptual background and usage is proposed in
a vignette (see \code{vignette(topic = "flows")}).
}

