% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{nav}
\alias{nav}
\title{Commuters}
\description{
Data on commuters between Urban Areas of the Grand-Est Regions in 2011.
}
\details{
Variables: \cr
\itemize{
\item{i: Code of urban area of residence}
\item{namei: Name of the urban area of residence}
\item{wi: Number of active occupied persons in the urban area of residence}
\item{j: Code of urban area of work}
\item{namej: Name of the urban area of work}
\item{wj: Number of active occupied persons in the urban area of work}
\item{fij: Number of persons commuting}
}
}
\references{
\url{http://www.insee.fr/fr/themes/detail.asp?reg_id=99&ref_id=mobilite-professionnelle-11}
}

