% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{plotMapDomFlows}
\alias{plotMapDomFlows}
\title{Dominant Flows Map}
\usage{
plotMapDomFlows(mat, spdf, spdfid, w, wid, wvar, wcex = 0.05,
  legend.flows.pos = "topright", legend.flows.title = "flow intensity",
  legend.nodes.pos = "topleft", add = FALSE)
}
\arguments{
\item{mat}{A square matrix of dominant flows}

\item{spdf}{Spatial*DataFrame of units}

\item{spdfid}{Name of the 'id' variable in the spdf data.frame}

\item{w}{A data.frame which contains a weight variable for the units}

\item{wid}{Identifier column  in w}

\item{wvar}{Weight variable}

\item{wcex}{Size factor for circles of weight}

\item{legend.flows.pos}{position of the flows legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{legend.flows.title}{title of the flows legend.}

\item{legend.nodes.pos}{position of the nodes legend, one of "topleft", "top",
"topright", "left", "right", "bottomleft", "bottom", "bottomright".}

\item{add}{Whether to add the plot to an existing one or not.}
}
\description{
Display a dominant flows map
}
\examples{
data(LoireAtlantique)
mat <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
diag(mat) <- 0
x <- domflows(mat = mat, wi = colSums(mat), wj = colSums(mat), k = 1)
firstx <- firstflows(mat = mat, method = "nfirst", ties.method = "first", k = 1)
hab <- mat * firstx * x
inflows <- data.frame(id = colnames(mat), w = colSums(mat))
par(mar = c(0,0,2,0))
sp::plot(COM44, col = "#cceae7")
plotMapDomFlows(mat = hab,
                spdf = COM44,
                spdfid = "INSEE_COM",
                w = inflows,
                wid = "id",
                wvar = "w",
                wcex = 0.05,
                add = TRUE,
                legend.flows.pos = "bottomleft",
                legend.flows.title = "Residential flows\\nintensity")
title("Dominant Residential Flows")
mtext(text = "INSEE, 2008", side = 4, line = -1, adj = 0.01, cex = 0.8)
}

