% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow.R
\name{statmat}
\alias{statmat}
\title{Descriptive Statistics of Flows Matrix}
\usage{
statmat(mat)
}
\arguments{
\item{mat}{A flow matrix}
}
\value{
The function returns graphics and statistics (invisible). \cr
\itemize{
\item{ nblinks: number of cells with values > 0,}
\item{ density: nbcellfull/nbcell}
\item{ sumflows: sum of flows}
\item{ min: min flow }
\item{ Q1: Q1 flow}
\item{ median: median fow}
\item{ Q3: Q3 flow}
\item{ max: max flow}
\item{ mean: mean flow}
\item{ sd: standart deviation flow}}
}
\description{
Give various indicators and graphical outputs on flow matrices
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
x <- statmat(myflows$mat)
x
}

