% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\name{statmat}
\alias{statmat}
\title{Descriptive Statistics of Flows Matrix}
\usage{
statmat(mat, output = "all")
}
\arguments{
\item{mat}{A square flow matrix.}

\item{output}{Type of graphical output. Choices are "all" for all graphics,
"none" to avoid graphical outputs, "degree" for degree distribution, "wdegree" for
weighted degree distribution, "lorenz" for Lorenz curve of link weights and
"boxplot" for boxplot of link weights}
}
\value{
The function returns graphics, statistics and a list. \cr
\itemize{
\item{nblinks: number of cells with values > 0,}
\item{density: number of links / number of possible links (also called gamma index by geographers)}
\item{connectcomp: number of connected components (isolates included,
weakly connected see \code{\link{clusters}})}
\item{connectcompx: number of connected components (isolates deleted,
weakly connected see \code{\link{clusters}})}
\item{sizecomp: a data.frame only returned in the list, size of the connected components.}
\item{compocomp: a data.frame only returned in the list, connected components membership of units.}
\item{sumflows: sum of flows}
\item{min: min flow }
\item{Q1: Q1 flow}
\item{median: median flow}
\item{Q3: Q3 flow}
\item{max: max flow}
\item{mean: mean flow}
\item{sd: standart deviation flow}}
Graphics returned are: \cr
\itemize{
\item{degree distribution}
\item{weigthed degree distribution}
\item{Lorenz curve regarding links weights}
\item{Boxplot regarding links weights}
}
}
\description{
Give various indicators and graphical outputs on flow matrices
}
\examples{
data(LoireAtlantique)
myflows <- prepflows(mat = MRE44, i = "DCRAN", j = "CODGEO", fij = "NBFLUX_C08_POP05P")
x <- statmat(myflows)
# Size of connected components
x$sizecomp
# Sum of flows
x$sumflows
x <- statmat(myflows)
# Size of connected components
x$sizecomp
# Sum of flows
x$sumflows
}

