% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{UA}
\alias{UA}
\title{Urban Areas}
\description{
SpatialPolygonsDataFrame of Urban Areas of the Grand Est region in France.
(2010 delineation).
}
\examples{
## UA
data(nav)
sp::plot(GE, col = "#cceae7", border = "grey50")
sp::plot(UA, col = "#940000", border = "white", add = TRUE)
}
\references{
\url{http://professionnels.ign.fr/geofla#tab-3}
}

